Rename InlineEditor to InstantEditor
This is the name already used in the documentation.
