+++ gaphor/UML/classes/classeseditors.py
@@ -1,13 +1,13 @@
 from gaphas.geometry import Rectangle, distance_point_point_fast
 
 from gaphor.core.format import format, parse
-from gaphor.diagram.inlineeditors import InlineEditor, popup_entry, show_popover
+from gaphor.diagram.instanteditors import InstantEditor, popup_entry, show_popover
 from gaphor.transaction import Transaction
 from gaphor.UML.classes.association import AssociationItem
 
 
-@InlineEditor.register(AssociationItem)
-def association_item_inline_editor(item, view, event_manager, pos=None) -> bool:
+@InstantEditor.register(AssociationItem)
+def association_item_editor(item, view, event_manager, pos=None) -> bool:
     """Text edit support for Named items."""
 
     subject = item.subject
